+++ app/src/main/kotlin/com/wire/android/ui/userprofile/other/OtherUserProfile.kt
@@ -1,0 +1,4 @@
+package com.wire.android.ui.userprofile.other
+
+class OtherUserProfile {
+}
